cosplay-simple/requirements.txt
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
pillow>=9.0.0
requests>=2.28.0